+++ deegree-services/deegree-webservices/src/main/java/org/deegree/console/webservices/MainConfig.java
@@ -73,7 +73,9 @@ public class MainConfig extends Config implements Serializable {
     @Override
     public String edit()
                             throws IOException {
-        FileUtils.copyURLToFile( getTemplate(), file );
+        if ( !file.exists() ) {
+            FileUtils.copyURLToFile( getTemplate(), file );
+        }
         StringBuilder sb = new StringBuilder( "/console/generic/xmleditor?faces-redirect=true" );
         sb.append( "&schemaUrl=" ).append( MAIN_SCHEMA_URL.toString() );
         sb.append( "&fileName=" ).append( workspaceName + "/" + FILE_NAME );
